Introducing new base form for changing app credential configurations.
